Add a new interface and a new java file
A problem with previous build is that after setting the timer, user cannot see the timer counting down, which means they do not know how much time they had left. To solve this problem, another interace called countdown.xml was added, which shows how many hours, minutes, and seconds there have left. A Relative Layout was used.

After creating an interface, setText() methods are used to set the time shown in countdown.xml to be the same as previous user input. Afterwards, a total amount of seconds are calculated based on how many hours, minutes, and seconds there are, storing the result in a variable. Finally, that variable is used in a CountDownTimer method to deduce how much time to wait before finishing.

A limitation is that after time is up, the message is not shown properly.