Add message if user does not enter a value
In previous build, if user does not enter an input in prompt interface, the app crashes. In this build, if...else statements are used to check whether the time values user have set are valid. If either hours, minutes and seconds section is blank, a message will pop up that prompts user to try again. Similarly, if all sections are 0, a message will pop up that states a timer consists of 0 seconds are not possible.

A limitation of this build is that the buttons in countdown interface is still not functional.