Made three buttons functional
In previous build, when timer is counting down, there are no buttons that allows users to control the timer. In this build, 3 buttons are created and made functional:

Pause button temporarily stops the timer from counting down. This is done by cancel() method in timer class, which terminates the timer and stops the value from descending.

Resume button creates a new instance of CountDownTimer, which adjusts its total time corresponding to the amount of time shown when user clicked pause. For example, if user initially sets timer to 20 seconds, but pauses the timer at 8, clicking resume means timer going for 8 seconds before stopping. This allows users to continue

New time button is designed so that if user wants to set the timer again, they are allowed to go back to the previous screen and reset new timer. This is done by first using cancel() method to first terminate original CountDownTimer(), and use getString() method to read user's input in setting screen.

A limitation of this build is that the "reset" button is still not functional