Add screenshots to what app does
In this build, screenshots are added to demonstrate what this app looks like. A screenshot of the homepage interface was added, where users can set how many hours, minutes, and seconds there are in timer. Another screenshot of countdown interface was added, where users can see the progress of the timer, and the message that it shows when timer has ran out.